backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
django-csp>=4.0
channels>=4.0
daphne>=4.0
Pillow>=10.0
python-dotenv>=1.0
dj-database-url>=2.0
whitenoise>=6.0
# Production uchun:
channels-redis>=4.0
redis>=5.0
django-redis>=5.4
psycopg2-binary>=2.9
django-storages==1.14.2
boto3==1.34.0
botocore==1.34.0
